fix: raise a BrokenPipeError only if the process exited nicely
On a BrokenPipeError we still always want an exception, but if the process
exited with a non-zero exit code, it's probably more helpful to show its
standard error, which is via a IterableSubprocessError. So in this case, we
swallow the BrokenPipeError and allow the code to fall through to the part that
raises the IterableSubprocessError.