feat: deal better with edge cases and race conditions
The main issue this addresses: a deadlock could happen if a KeyboardInterrupt
is raised due to the stderr thread reading from stderr, and the main thread
trying to close it at the same time. Now this is (hopefully) impossible, at
least for well-behaved processes, since the threads are only created in a
try/except block that catches such exceptions, and terminates the process.

Potentially this doesn't solve the issue if the process doesn't respond to
termination, but leaving that investigation to another time. Suspect this is
still a good step forward.

There are some additional tests on related behaviour on exit of the context.
Have to admit, not completely sure what is desirable. But they at least
document what is done now, so if we do want things to change, they will be
deliberate.